JDBC Push
Started working on JDBC source. Finished creating the Tables for Resources and Race. I can make a Skills Table if desired as well. The scripts for those tables are in the lib/dnddb folder.
Also added and referenced mysql-connector.jar, which from what I read is required for interfacing mysql with jdatabase.
Lastly, I added a Jdbc.java class that handles all of the database interactions. There's still stuff left to do with it and it hasn't been tested:
Figure out a way choose which Table to access so as to not have to copy-and-paste each method to handle each table individually.
Create a removeFromTable() method.
Test the methods.
Figure out how this will interact with the objects.
Determine if it would be fine to just have this be exported and used by another computer, or if extra steps need to be taken so that the database can be remotely accessed. This is because right now I believe I'm just accessing it from a local host.